STT/requirements.txt
Flask
google-cloud-speech
transformers
pydub
python-dotenv
requests
Flask-Cors

